prep redo flushing
