+++ crates/polars-pipe/src/executors/sinks/joins/mod.rs
@@ -78,7 +78,7 @@ impl Default for Tracker {
     #[inline]
     fn default() -> Self {
         Self {
-            inner: Default::default(),
+            inner: AtomicBool::new(false)
         }
     }
 }
